Add files via upload
Excel-ERP provides filling Uyumsoft ERP software automatically for your ordering of material about design projects. When you copy and paste materials for ordering from Solidworks software in the first page of excel and click the button, it'll be arranged worksheet 2 first, then it's ok when you drag an drop the file into te Uyumsoft ERP software but you need to order column first according to worksheet 2 in excel.

kırılan kalıplar provide you a form to prepare report. You can fullfill the form and click on the "Ekle", it'll be added at last in the excel.